requirements.txt
Django>=5.0,<6.0
djangorestframework>=3.15
# simplejwt v5 is compatible with DRF; v8 doesn't exist. Pin to >=5,<6
djangorestframework-simplejwt>=5.0,<6.0
django-cors-headers>=4.0
# Use latest compatible django-mssql-backend available on PyPI
django-mssql-backend>=2.8.1,<3.0
pyodbc>=4.0
reportlab>=4.0
python-dotenv>=1.0
Pillow>=10.0